Alterações na base de dados

+++ HockeyPT/HockeyPT/Models/Jogadores.cs
@@ -24,9 +24,9 @@ namespace HockeyPT.Models
 
         public string Nacionalidade { get; set; }
 
-        public double Altura { get; set; }
+        public int Altura { get; set; }
 
-        public double Peso { get; set; }
+        public int Peso { get; set; }
 
         public string Fotografia { get; set; }
